Admin이 아닌 고객과 관련된 메서드 위치변경 ( ItemRepository -> ItemCustomerRepository )
